add backend service tests
